+update load and run stim to take separate addrs for a1 and a2 +fix bug in append nop vec in dots to prepend V char since expand dots expects it +add writes, reads, load, loads, loada, unload, unload-all, run, runc, set-profile, fail-pins and get-pins funcs to prgm +change get_stim_by_path to take a profile pointer instead of a profile path string

+++ lib/fe/fe.h
@@ -17,7 +17,7 @@ extern "C" {
 
 #define FE_VERSION "1.0"
 
-typedef float fe_Number;
+typedef double fe_Number;
 typedef struct fe_Object fe_Object;
 typedef struct fe_Context fe_Context;
 typedef fe_Object* (*fe_CFunc)(fe_Context *ctx, fe_Object *args);
